+++ sources/input_output/move.c
@@ -6,7 +6,7 @@
 /*   By: lollith <lollith@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/21 16:23:02 by agouet            #+#    #+#             */
-/*   Updated: 2022/12/11 10:17:07 by lollith          ###   ########.fr       */
+/*   Updated: 2022/12/11 14:27:41 by lollith          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,19 +54,19 @@ void	move(double new_pos_x, double new_pos_y, t_all *all, int sign)
 	check_collision(all, old_x, old_y, sign);
 }
 
-void	rotate(t_all *all, int sign)
+void	rotate(t_all *all, int sign, double speed)
 {
 	double	old_orient_x;
 	double	old_plane_x;
 
 	old_orient_x = all->ray.orient_x;
 	old_plane_x = all->ray.plane_x;
-	all->ray.orient_x = all->ray.orient_x * cos(sign * MOVE_SPEED)
-		- all->ray.orient_y * sin(sign * MOVE_SPEED);
-	all->ray.plane_x = all->ray.plane_x * cos(sign * MOVE_SPEED)
-		- all->ray.plane_y * sin (sign * MOVE_SPEED);
-	all->ray.orient_y = old_orient_x * sin(sign * MOVE_SPEED)
-		+ all->ray.orient_y * cos(sign * MOVE_SPEED);
-	all->ray.plane_y = old_plane_x * sin(sign * MOVE_SPEED)
-		+ all->ray.plane_y * cos (sign * MOVE_SPEED);
+	all->ray.orient_x = all->ray.orient_x * cos(sign * speed)
+		- all->ray.orient_y * sin(sign * speed);
+	all->ray.plane_x = all->ray.plane_x * cos(sign * speed)
+		- all->ray.plane_y * sin (sign * speed);
+	all->ray.orient_y = old_orient_x * sin(sign * speed)
+		+ all->ray.orient_y * cos(sign * speed);
+	all->ray.plane_y = old_plane_x * sin(sign * speed)
+		+ all->ray.plane_y * cos (sign * speed);
 }
